+++ src/yaplox/stmt.py
@@ -20,11 +20,19 @@ class StmtVisitor(ABC):
         raise NotImplementedError
 
     @abstractmethod
+    def visit_function_stmt(self, stmt: "Function"):
+        raise NotImplementedError
+
+    @abstractmethod
     def visit_if_stmt(self, stmt: "If"):
         raise NotImplementedError
 
     @abstractmethod
     def visit_print_stmt(self, stmt: "Print"):
+        raise NotImplementedError
+
+    @abstractmethod
+    def visit_return_stmt(self, stmt: "Return"):
         raise NotImplementedError
 
     @abstractmethod
@@ -60,6 +68,17 @@ class Expression(Stmt):
         return visitor.visit_expression_stmt(self)
 
 
+class Function(Stmt):
+    def __init__(self, name: Token, params: List[Token], body: List[Stmt]):
+        self.name = name
+        self.params = params
+        self.body = body
+
+    def accept(self, visitor: StmtVisitor):
+        """ Create a accept method that calls the visitor. """
+        return visitor.visit_function_stmt(self)
+
+
 class If(Stmt):
     def __init__(self, condition: Expr, then_branch: Stmt, else_branch: Optional[Stmt]):
         self.condition = condition
@@ -78,6 +97,16 @@ class Print(Stmt):
     def accept(self, visitor: StmtVisitor):
         """ Create a accept method that calls the visitor. """
         return visitor.visit_print_stmt(self)
+
+
+class Return(Stmt):
+    def __init__(self, keyword: Token, value: Optional[Expr]):
+        self.keyword = keyword
+        self.value = value
+
+    def accept(self, visitor: StmtVisitor):
+        """ Create a accept method that calls the visitor. """
+        return visitor.visit_return_stmt(self)
 
 
 class Var(Stmt):